refactor: :recycle: adjustments for extracting codemeta.json info
